[Mirror] Integrated newest Mirror reflection library changes
Updating data to match new Mirror formatting. QwerkE compiles and runs, although entity script updates aren't called but that's fine for now. A big step to improving de/serialization

+++ Source/Editor/QE_EditorWindowMenuBar.h
@@ -132,7 +132,7 @@ namespace QwerkE {
                     }
                     else
                     {
-                        ImGui::SameLineCentered(currentProject.projectName.size());
+                        ImGui::SameLineCentered((u16)currentProject.projectName.size());
                         ImGui::Text(currentProject.projectName.c_str());
                     }
